[ci skip] refactor: Separate DrawTask and DrawCardTask
