+++ Sources/hspp/Tasks/SimpleTasks/DrawTask.cpp
@@ -5,9 +5,6 @@
 
 #include <hspp/Commons/Constants.hpp>
 #include <hspp/Models/Entity.hpp>
-#include <hspp/Models/Game.hpp>
-#include <hspp/Models/Minion.hpp>
-#include <hspp/Models/Weapon.hpp>
 #include <hspp/Policies/Policy.hpp>
 #include <hspp/Tasks/SimpleTasks/DrawTask.hpp>
 
@@ -94,49 +91,5 @@ TaskStatus DrawTask::Impl(Player& player)
     }
 
     return result;
-}
-
-DrawCardTask::DrawCardTask(Card card) : m_card(std::move(card))
-{
-    // Do nothing
-}
-
-TaskID DrawCardTask::GetTaskID() const
-{
-    return TaskID::DRAW;
-}
-
-TaskStatus DrawCardTask::Impl(Player& player)
-{
-    //auto& deck = player.GetDeck();
-    auto& hand = player.GetHand();
-
-    auto* game = player.GetGame();
-
-    switch (m_card.cardType)
-    {
-        case +CardType::MINION:
-        {
-            const auto minion = new Minion(game, m_card);
-            hand.AddCard(*minion);
-            break;
-        }
-        case +CardType::WEAPON:
-        {
-            const auto weapon = new Weapon(game, m_card);
-            hand.AddCard(*weapon);
-            break;
-        }
-        default:
-            throw std::invalid_argument(
-                "DrawCardTask::Impl() - Invalid card type!");
-    }
-
-    //if (!deck.empty())
-    //{
-    //    deck.pop_back();
-    //}
-
-    return TaskStatus::DRAW_SUCCESS;
 }
 }  // namespace Hearthstonepp::SimpleTasks
